Add files via upload
I think the problem happens when adding multiple blocks, because promises are async, _addBlock keeps on adding new blocks to the blockchain before it has resolved from adding previous blocks. So, they all get added as if they were genesis! I tried fixing the problem but it only works for the first block, the second and third blocks are added as if they were still the first block!

The question is, I can move on from this part if there is a certain delay that will always happen between the adding of new blocks to give a chance for the promise to resolve and the blockchain to be updated...